SDL-T-64: Add new dependencies and update ignored files
This commit adds new dependencies to the `Cargo.toml` file, including `axum`, `serde_json`, and `serde_with`. It also modifies the list of ignored files in the `.gitignore` file to include the `.fleet/` and `.idea/` directories. Additionally, the `docker-compose.yaml` file is modified to use quotation marks for the port mapping. In the `src/db/mod.rs` file, the `repository` module is made public, and in the `src/main.rs` file, the `model`, `rest_api`, and `time` modules are imported. The `src/db/repositiry.rs` file is renamed to `src/model/mod.rs`. Finally, a new file `src/rest_api/mod.rs` is added, which includes a skeleton function for handling API routes.